полностью работает под линукс
git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@63690 954022d7-b5bf-4e40-9824-e11837661b57

+++ HtmlRenderer/src/Text.h
@@ -914,8 +914,8 @@ namespace NSHtmlRenderer
 
             if (nCount > 1)
             {
-                LONG* pWidthBuf = (LONG*)(m_pPageMeta->GetData() + _position);
-                *pWidthBuf = (LONG)(dWidthLine * 10000);
+                int* pWidthBuf = (int*)(m_pPageMeta->GetData() + _position);
+                *pWidthBuf = (int)(dWidthLine * 10000);
             }
 
             m_oLine.Clear();
